+++ changelog.md
@@ -10,6 +10,8 @@ If you are reading this on the website, then these changes are live here.
 - [`each` `∵`](https://uiua.org/docs/each) and [`rows` `≡`](https://uiua.org/docs/rows) now work with [`under` `⍜`](https://uiua.org/docs/under)
 - Parsing multiple formattable functions from words is now smarter
 - Remove `bind '`. It made code hard to read. It will continue to parse, but will be formatted as `(…)`
+### Interpreter
+- Add the `uiua stand` command, which creates a standalone executable
 
 ## 0.0.25 - 2023-10-29
 ### Interpreter